sdk/go: simplify source code, remove dead code and duplication
Remove unused proto-to-domain converters (authenticatedUserFromProto,
storedUserFromProto, protoKindToProvider) and a dead time.Time branch in
catalogParameterType. Simplify catalogToJSON by replacing the parallel
wireParameter/wireOperation/wireCatalog type hierarchy with struct
embedding, eliminating ~8 dead wire fields. Remove double panic recovery
in the Execute path, fix for-as-if in isOptionalType, use os.Remove
instead of os.RemoveAll for socket cleanup, and remove nil-receiver
guards on operationError.

Consolidate scattered utilities into convert.go (cloneStringMap, time
helpers, configFromRequest), merge runtime.go into types.go and
errors.go with operation_errors.go, and fix make+append patterns in
datastore list methods. Net result is 125 fewer lines across 11 files
with no public API changes.
